requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cron-descriptor==1.4.0
cycler==0.12.1
Django==4.2.8
django-autocomplete-light==3.3.2
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-filter==23.5
django-haystack==2.8.1
django-haystack-elasticsearch5==0.7
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
django-tastypie==0.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.46.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
inflection==0.5.1
kiwisolver==1.4.5
kombu==5.3.4
networkx==3.2.1
numpy==1.26.2
packaging==23.2
phonenumbers==8.13.27
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.11.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
